[test] (all) Fix failing tests
