+++ packages/coinlib-doge/test/bip44.test.ts
@@ -1,7 +1,6 @@
 import {
   deriveAddress,
   derivePrivateKey,
-  splitDerivationPath,
   deriveHDNode,
   deriveKeyPair,
   xprvToXpub,
@@ -18,11 +17,6 @@ import {
 export const BASE_NODE = bip32.fromBase58(DERIVED_XPRV, NETWORK_MAINNET)
 
 describe('bip44', () => {
-  describe('splitDerivationPath', () => {
-    it('returns correct value', () => {
-      expect(splitDerivationPath(DERIVATION_PATH)).toEqual(["44'", "3'", "0'"])
-    })
-  })
   describe('deriveHDNode', () => {
     it('derives root xprv correctly', () => {
       expect(deriveHDNode(ROOT_XPRV, DERIVATION_PATH, NETWORK).toBase58()).toEqual(DERIVED_XPRV)
