+++ packages/coinlib/test/utils.test.ts
@@ -1,14 +1,7 @@
-import { keysOf, getSeedHexFromMnemonic, getFingerprintFromSeedHex } from '../src/utils'
+import { getSeedHexFromMnemonic, getFingerprintFromSeedHex } from '../src/utils'
 import { MNEMONIC, SEED_HEX, FINGERPRINT } from './fixtures/utils'
 
 describe('utils', () => {
-  describe('keysOf', () => {
-    it('works', () => {
-      const result: ('a' | 'b' | 'c')[] = keysOf({ a: 1, b: 2, c: 5 })
-      expect(result).toEqual(['a', 'b', 'c'])
-    })
-  })
-
   describe('getSeedHexFromMnemonic', () => {
     it('works', () => {
       const seedHex = getSeedHexFromMnemonic(MNEMONIC)
